+++ .tool-versions
@@ -1,0 +1,1 @@
+ruby 2.7.3
