+++ spec/support/invoke.rb
@@ -1,0 +1,4 @@
+def invoke(command)
+  Timetrap::CLI.parse command
+  Timetrap::CLI.invoke
+end
